Modified the path at which the lecture files are saved.

+++ src/com/google/gwt/sample/elearning/server/service/LectureUploadService.java
@@ -19,6 +19,7 @@ import java.util.Iterator;
 import java.util.List;
 
 /**
+ *
  * Created by razvanolar on 28.11.2015.
  */
 public class LectureUploadService extends HttpServlet {
@@ -27,29 +28,29 @@ public class LectureUploadService extends HttpServlet {
   protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
     FileItemFactory factory = new DiskFileItemFactory();
     ServletFileUpload upload = new ServletFileUpload(factory);
-    String user = req.getParameter("user");
     String lecture = req.getParameter("lecture");
     String path = req.getParameter("path");
     try {
-      long userId = Long.parseLong(user);
       long lectureId = Long.parseLong(lecture);
-      String filePath = ServerUtil.getUserLectureDirectoryPath(userId, lectureId) + path + "\\";
+      path = path != null && !path.isEmpty() ? path + "\\" : path;
+      String filePath = ServerUtil.getUserLectureDirectoryPath(lectureId) + path;
+      File dirs = new File(filePath);
+      if (!dirs.exists() && !dirs.mkdirs())
+        throw new Exception("Unable to create missing directories");
       List<FileItem> items = upload.parseRequest(req);
       Iterator<FileItem> iterator = items.iterator();
       while (iterator.hasNext()) {
         FileItem item = iterator.next();
         if (item.isFormField()) {
-          String name = item.getFieldName();
-          String value = item.getString();
           System.out.println();
         } else {
           String fieldName = item.getFieldName();
-          String fileName = item.getName();
-          if (fieldName != null)
-            fileName = FilenameUtils.getName(fileName);
-          String contentType = item.getContentType();
-          boolean isInMemory = item.isInMemory();
-          long sizeInBytes = item.getSize();
+//          String fileName = item.getName();
+//          if (fieldName != null)
+//            fileName = FilenameUtils.getName(fileName);
+//          String contentType = item.getContentType();
+//          boolean isInMemory = item.isInMemory();
+//          long sizeInBytes = item.getSize();
           byte[] data = item.get();
           FileOutputStream fileOutSt = new FileOutputStream(filePath + item.getName());
           fileOutSt.write(data);
